Edit migration "create_evaluations_table" to remove default value
The column "template_name" is unique, however its default value is
"NULL" (as a string), which would prevent more than one entry to be
inserted with that default value. As Evaluations should be
able to not have a template_name (if they're not a template but an
actual Evaluation meant to be filled), we make the default value of
template_name NULL (the actual SQL value), allowing multiple Evaluations
to not have a template_name, as MySQL doesnt check uniqueness on fields
having NULL as their value.